+++ tests/test_download_photos.py
@@ -2165,3 +2165,41 @@ class DownloadPhotoTestCase(TestCase):
 
         self.assertIsInstance(result.exception, ValueError)
         # self.assertEqual(result.exception, ValueError("Invalid Input: b'i\\xb7\\x1dy\\xf8!'"))
+
+    def test_download_filename_string_encoding(self) -> None:
+        base_dir = os.path.join(self.fixtures_path, inspect.stack()[0][3])
+
+        files_to_create = [
+            ("2018/07/30", "IMG_7408.JPG", 1151066),
+            ("2018/07/30", "IMG_7407.JPG", 656257),
+        ]
+
+        files_to_download: List[Tuple[str, str]] = [
+            # SU1HXzc0MDkuSlBH -> IMG_7409.JPG
+            ("2018/07/31", "IMG_7409.JPG")
+        ]
+
+        data_dir, result = run_icloudpd_test(
+            self.assertEqual,
+            self.vcr_path,
+            base_dir,
+            "listing_photos_filename_string_encoding.yml",
+            files_to_create,
+            files_to_download,
+            [
+                "--username",
+                "jdoe@gmail.com",
+                "--password",
+                "password1",
+                "--recent",
+                "5",
+                "--skip-videos",
+                "--skip-live-photos",
+                "--no-progress-bar",
+                "--threads-num",
+                "1",
+            ],
+        )
+        print_result_exception(result)
+
+        self.assertEqual(result.exit_code, 0)